amplify/agent/requirements.txt
botocore
bedrock-agentcore
requests
pytz

strands-agents
strands-agents[otel]
strands-agents-tools[rss]
